Change name of prebuild to fix erroneous script execution
When the build command was run, it was supposed to trigger download, process and prebuild in sequence (&&). That wasn't happening. Prebuild was called first and that errored out b/c there weren't any data files (duh!). I rewrote the npm scripts as "pre-build" instead of "prebuild" and now the scripts run in proper sequence. I don't really know why that worked but it seems to be a temp cure. Needs more investigation.